+++ example/lib/screens/component/apple_pay/apple_pay_session_component_screen.dart
@@ -3,6 +3,7 @@
 import 'package:adyen_checkout/adyen_checkout.dart';
 import 'package:adyen_checkout_example/config.dart';
 import 'package:adyen_checkout_example/repositories/adyen_apple_pay_component_repository.dart';
+import 'package:adyen_checkout_example/utils/dialog_builder.dart';
 import 'package:flutter/material.dart';
 
 class ApplePaySessionComponentScreen extends StatelessWidget {
@@ -67,7 +68,7 @@ class ApplePaySessionComponentScreen extends StatelessWidget {
                 height: 48,
                 onPaymentResult: (paymentResult) {
                   Navigator.pop(context);
-                  _dialogBuilder(paymentResult, context);
+                  DialogBuilder.showPaymentResultDialog(paymentResult, context);
                 },
               ),
             ],
@@ -105,46 +106,6 @@ class ApplePaySessionComponentScreen extends StatelessWidget {
     return paymentMethods["paymentMethods"].firstWhere(
       (paymentMethod) => paymentMethod["type"] == "applepay",
       orElse: () => throw Exception("Apple pay payment method not provided"),
-    );
-  }
-
-  _dialogBuilder(PaymentResult paymentResult, BuildContext context) {
-    String title = "";
-    String message = "";
-    switch (paymentResult) {
-      case PaymentAdvancedFinished():
-        title = "Finished";
-        message = "Result code: ${paymentResult.resultCode}";
-      case PaymentSessionFinished():
-        title = "Finished";
-        message = "Result code: ${paymentResult.resultCode}";
-      case PaymentCancelledByUser():
-        title = "Cancelled by user";
-        message = "Cancelled by user";
-      case PaymentError():
-        title = "Error occurred";
-        message = "${paymentResult.reason}";
-    }
-
-    return showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text(title),
-          content: Text(message),
-          actions: <Widget>[
-            TextButton(
-              style: TextButton.styleFrom(
-                textStyle: Theme.of(context).textTheme.labelLarge,
-              ),
-              child: const Text('Close'),
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-            ),
-          ],
-        );
-      },
     );
   }
 }