+++ example/lib/screens/component/card/card_component_scrollable_screen.dart
@@ -3,6 +3,7 @@
 import 'package:adyen_checkout/adyen_checkout.dart';
 import 'package:adyen_checkout_example/config.dart';
 import 'package:adyen_checkout_example/repositories/adyen_card_component_repository.dart';
+import 'package:adyen_checkout_example/utils/dialog_builder.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/material.dart';
 
@@ -72,7 +73,7 @@ class CardComponentScrollableScreen extends StatelessWidget {
       checkout: advancedCheckout,
       onPaymentResult: (paymentResult) async {
         Navigator.pop(context);
-        _dialogBuilder(context, paymentResult);
+        DialogBuilder.showPaymentResultDialog(paymentResult, context);
       },
     );
   }
@@ -91,43 +92,5 @@ class CardComponentScrollableScreen extends StatelessWidget {
         storedPaymentMethodList.firstOrNull;
 
     return paymentMethod ?? <String, String>{};
-  }
-
-  _dialogBuilder(BuildContext context, PaymentResult paymentResult) {
-    String title = "";
-    String message = "";
-    switch (paymentResult) {
-      case PaymentAdvancedFinished():
-        title = "Finished";
-        message = "Result code: ${paymentResult.resultCode}";
-      case PaymentSessionFinished():
-        title = "Finished";
-        message = "Result code: ${paymentResult.resultCode}";
-      case PaymentError():
-        title = "Error occurred";
-        message = "${paymentResult.reason}";
-      default:
-    }
-
-    return showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text(title),
-          content: Text(message),
-          actions: <Widget>[
-            TextButton(
-              style: TextButton.styleFrom(
-                textStyle: Theme.of(context).textTheme.labelLarge,
-              ),
-              child: const Text('Close'),
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-            ),
-          ],
-        );
-      },
-    );
   }
 }