+++ example/lib/screens/component/instant/instant_session_component_screen.dart
@@ -3,6 +3,7 @@
 import 'package:adyen_checkout/adyen_checkout.dart';
 import 'package:adyen_checkout_example/config.dart';
 import 'package:adyen_checkout_example/repositories/adyen_instant_component_repository.dart';
+import 'package:adyen_checkout_example/utils/dialog_builder.dart';
 import 'package:flutter/material.dart';
 
 class InstantSessionComponentScreen extends StatelessWidget {
@@ -66,7 +67,8 @@ class InstantSessionComponentScreen extends StatelessWidget {
                           checkout: sessionCheckout,
                         )
                         .then((paymentResult) =>
-                            _dialogBuilder(paymentResult, context));
+                            DialogBuilder.showPaymentResultDialog(
+                                paymentResult, context));
                   },
                   child: const Text("Paypal")),
               TextButton(
@@ -78,7 +80,8 @@ class InstantSessionComponentScreen extends StatelessWidget {
                           checkout: sessionCheckout,
                         )
                         .then((paymentResult) =>
-                            _dialogBuilder(paymentResult, context));
+                            DialogBuilder.showPaymentResultDialog(
+                                paymentResult, context));
                   },
                   child: const Text("Klarna"))
             ],
@@ -95,46 +98,6 @@ class InstantSessionComponentScreen extends StatelessWidget {
     return paymentMethods["paymentMethods"].firstWhere(
       (paymentMethod) => paymentMethod["type"] == key,
       orElse: () => throw Exception("$key payment method not provided"),
-    );
-  }
-
-  _dialogBuilder(PaymentResult paymentResult, BuildContext context) {
-    String title = "";
-    String message = "";
-    switch (paymentResult) {
-      case PaymentAdvancedFinished():
-        title = "Finished";
-        message = "Result code: ${paymentResult.resultCode}";
-      case PaymentSessionFinished():
-        title = "Finished";
-        message = "Result code: ${paymentResult.resultCode}";
-      case PaymentCancelledByUser():
-        title = "Cancelled by user";
-        message = "Cancelled by user";
-      case PaymentError():
-        title = "Error occurred";
-        message = "${paymentResult.reason}";
-    }
-
-    return showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text(title),
-          content: Text(message),
-          actions: <Widget>[
-            TextButton(
-              style: TextButton.styleFrom(
-                textStyle: Theme.of(context).textTheme.labelLarge,
-              ),
-              child: const Text('Close'),
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-            ),
-          ],
-        );
-      },
     );
   }
 }