+++ example/lib/screens/component/multi_component/multi_component_advanced_screen.dart
@@ -5,6 +5,7 @@ import 'package:adyen_checkout_example/config.dart';
 import 'package:adyen_checkout_example/repositories/adyen_apple_pay_component_repository.dart';
 import 'package:adyen_checkout_example/repositories/adyen_card_component_repository.dart';
 import 'package:adyen_checkout_example/repositories/adyen_google_pay_component_repository.dart';
+import 'package:adyen_checkout_example/utils/dialog_builder.dart';
 import 'package:collection/collection.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
@@ -81,7 +82,7 @@ class MultiComponentAdvancedScreen extends StatelessWidget {
         checkout: advancedCheckout,
         onPaymentResult: (paymentResult) async {
           Navigator.pop(context);
-          _dialogBuilder(context, paymentResult);
+          DialogBuilder.showPaymentResultDialog(paymentResult, context);
         },
       ),
     );
@@ -138,7 +139,7 @@ class MultiComponentAdvancedScreen extends StatelessWidget {
         width: double.infinity,
         onPaymentResult: (paymentResult) {
           Navigator.pop(context);
-          _dialogBuilder(context, paymentResult);
+          DialogBuilder.showPaymentResultDialog(paymentResult, context);
         },
       ),
     );
@@ -178,7 +179,7 @@ class MultiComponentAdvancedScreen extends StatelessWidget {
         height: 56,
         onPaymentResult: (paymentResult) {
           Navigator.pop(context);
-          _dialogBuilder(context, paymentResult);
+          DialogBuilder.showPaymentResultDialog(paymentResult, context);
         },
       ),
     );
@@ -252,45 +253,5 @@ class MultiComponentAdvancedScreen extends StatelessWidget {
         storedPaymentMethodList.firstOrNull;
 
     return paymentMethod ?? <String, String>{};
-  }
-
-  _dialogBuilder(BuildContext context, PaymentResult paymentResult) {
-    String title = "";
-    String message = "";
-    switch (paymentResult) {
-      case PaymentAdvancedFinished():
-        title = "Finished";
-        message = "Result code: ${paymentResult.resultCode}";
-      case PaymentSessionFinished():
-        title = "Finished";
-        message = "Result code: ${paymentResult.resultCode}";
-      case PaymentError():
-        title = "Error occurred";
-        message = "${paymentResult.reason}";
-      case PaymentCancelledByUser():
-        title = "Cancelled by user";
-        message = "Cancelled by user";
-    }
-
-    return showDialog(
-      context: context,
-      builder: (BuildContext context) {
-        return AlertDialog(
-          title: Text(title),
-          content: Text(message),
-          actions: <Widget>[
-            TextButton(
-              style: TextButton.styleFrom(
-                textStyle: Theme.of(context).textTheme.labelLarge,
-              ),
-              child: const Text('Close'),
-              onPressed: () {
-                Navigator.of(context).pop();
-              },
-            ),
-          ],
-        );
-      },
-    );
   }
 }