+++ lib/src/common/model/payment_result.dart
@@ -1,9 +1,10 @@
 import 'package:adyen_checkout/src/common/model/order_response.dart';
+import 'package:adyen_checkout/src/common/model/result_code.dart';
 
 sealed class PaymentResult {}
 
 class PaymentAdvancedFinished extends PaymentResult {
-  final String resultCode;
+  final ResultCode resultCode;
 
   PaymentAdvancedFinished({required this.resultCode});
 }
@@ -11,7 +12,7 @@ class PaymentAdvancedFinished extends PaymentResult {
 class PaymentSessionFinished extends PaymentResult {
   final String sessionId;
   final String sessionData;
-  final String resultCode;
+  final ResultCode resultCode;
   final OrderResponse? order;
 
   PaymentSessionFinished({