+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.4 (in development)
 
+* ⚠ **Breaking Change**: Changed ResultCode into an enum. Please adjust your usage of the PaymentResult (PaymentSessionFinished or PaymentAdvancedFinished). 
 * ⚠ **Breaking Change**: The type of the paymentMethodsResponse for using Drop-in with the advanced flow changed to Map<String, dynamic>.  
 * Updated iOS SDK to v5.7.1.
 * Updated Android SDK to v5.3.1.